requirements.txt

piexif=1.1.3
pillow=9.1.0
